separate map/validate sidebar components in a different file
